handleDelete novo

+++ src/pages/InfoPage/InfoPage.js
@@ -26,20 +26,25 @@ function InfoPage (){
     }, []);
 
     function handleToast(){
-      toast((t) => (
-        <span>
-          <p>Are you sure you want to <b>delete</b> {strain.name}?</p>
-          <button onClick={() => toast.dismiss(t.id)} style={{marginRight:"0.5rem"}} className="btn btn-danger">
-            Nao
-          </button>
-          <button className= "btn btn-success" onClick={()=>{
-            handleDelete(t);
-            }}
-          >
-            Sim
-          </button>
-        </span>
-      ));
+      if(strain.isLocked === false){
+        toast((t) => (
+          <span>
+            <p>Are you sure you want to <b>delete</b> {strain.name}?</p>
+            <button onClick={() => toast.dismiss(t.id)} style={{marginRight:"0.5rem"}} className="btn btn-danger">
+              Nao
+            </button>
+            <button className= "btn btn-success" onClick={()=>{
+              handleDelete(t);
+              }}
+            >
+              Sim
+            </button>
+          </span>
+        ));
+      } else {
+        toast.error("This strain cannot be deleted")
+      }
+      
     }
 
     function handleEdit(){
